+++ test/object-store/sync/file.cpp
@@ -16,12 +16,13 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 
-#include "sync_test_utils.hpp"
-#include "util/test_utils.hpp"
-#include "util/test_file.hpp"
+#include <util/test_utils.hpp>
+#include <util/test_file.hpp>
+#include <util/sync/sync_test_utils.hpp>
 
 #include <realm/object-store/shared_realm.hpp>
 #include <realm/object-store/sync/sync_manager.hpp>
+
 #include <realm/util/file.hpp>
 
 using namespace realm;